% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdays.R
\name{ds_add_workdays}
\alias{ds_add_workdays}
\title{计算时间段内包含的工作日数量}
\usage{
ds_add_workdays(d, fromColumn, toColumn, newName = "@workdays")
}
\arguments{
\item{d}{要补充字段的数据集}

\item{fromColumn}{开始日期字段名}

\item{toColumn}{截止日期字段名}

\item{workdayName}{工作日字段命名}
}
\description{
计算数据集中指定两列占用的工作日数量
}
\seealso{
Other workday-funcs: 
\code{\link{ds_add_holidays}()},
\code{\link{holiday_assign}()},
\code{\link{holidays_zh_CN}()},
\code{\link{workday_assign}()},
\code{\link{workdays_zh_CN}()}
}
\concept{workday-funcs}
