% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_append}
\alias{ds_append}
\title{批量追加更新的数据}
\usage{
ds_append(d, dsName, topic = "CACHE", warn.ignore = F)
}
\arguments{
\item{d}{要写入的数据}

\item{dsName}{数据集名称}

\item{topic}{数据集保存的主题目录，默认为CACHE}

\item{toDelete}{将数据标记为删除}
}
\description{
为了避免修改原数据文件，追加数据时不查询旧数据，提交归档时再统一处理。

当追加的数据量与归档数据相比规模较小时，这种模式将具有优势。
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_as_deleted}()},
\code{\link{ds_as_from}()},
\code{\link{ds_collect}()},
\code{\link{ds_data}()},
\code{\link{ds_delete}()},
\code{\link{ds_drop}()},
\code{\link{ds_exists}()},
\code{\link{ds_files}()},
\code{\link{ds_init}()},
\code{\link{ds_read0}()},
\code{\link{ds_read_affected}()},
\code{\link{ds_read_appended}()},
\code{\link{ds_read_archived}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_parts}()},
\code{\link{ds_search}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
