% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-plyr.R
\name{ds_as_datetime}
\alias{ds_as_datetime}
\title{转换字符串为时间日期类型}
\usage{
ds_as_datetime(ds, timestampColumn, datetimeColumn, tzone = "Asia/Shanghai")
}
\description{
转换字符串为时间日期类型
}
\seealso{
Other ds function: 
\code{\link{ds_add_count}()},
\code{\link{ds_arrange}()},
\code{\link{ds_as_month}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_as_year}()},
\code{\link{ds_count}()},
\code{\link{ds_distinct}()},
\code{\link{ds_dup_count}()},
\code{\link{ds_filter}()},
\code{\link{ds_head}()},
\code{\link{ds_n_max}()},
\code{\link{ds_n_min}()},
\code{\link{ds_rename}()},
\code{\link{ds_select_by}()},
\code{\link{ds_select_na}()},
\code{\link{ds_select_not_na}()},
\code{\link{ds_select}()},
\code{\link{ds_tail}()}
}
\concept{ds function}
