% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-meta.R
\name{ds_as_schema}
\alias{ds_as_schema}
\title{按数据集架构定义转换数据框}
\usage{
ds_as_schema(ds, dsName, topic = "CACHE")
}
\arguments{
\item{ds}{数据框}

\item{dsName}{数据集名称}

\item{topic}{存储主题域}
}
\description{
查看数据集架构时，必须包含数据。
}
\seealso{
Other metadata function: 
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_meta_write}()},
\code{\link{ds_meta}()},
\code{\link{ds_schema}()}
}
\concept{metadata function}
