% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-meta.R
\name{ds_diff_schema}
\alias{ds_diff_schema}
\title{详细比较两个数据集结构}
\usage{
ds_diff_schema(schema1, schema2)
}
\arguments{
\item{schema1}{用于比较的数据集结构，可使用\code{\link{ds_schema}}获得}

\item{schema2}{用于参考的数据集结构，可使用\code{\link{ds_schema}}获得}
}
\description{
详细比较两个数据集结构
}
\seealso{
Other metadata function: 
\code{\link{ds_as_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_meta_write}()},
\code{\link{ds_meta}()},
\code{\link{ds_schema}()}
}
\concept{metadata function}
