% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_init}
\alias{ds_init}
\title{初始化数据集配置}
\usage{
ds_init(
  dsName,
  topic = "CACHE",
  data = tibble(),
  bySchema = list(),
  partColumns = c(),
  keyColumns = c(),
  suggestedColumns = c(),
  titleColumn = c(),
  drop = F,
  withData = F,
  desc = "-",
  type = "__UNKNOWN__"
)
}
\arguments{
\item{dsName}{数据集名称}

\item{topic}{数据集保存的主题目录，默认为CACHE}

\item{data}{样本数据，用来分析数据结构}

\item{partColumns}{支持分区存储，支持多列}

\item{keyColumns}{要求在所有行中具有唯一值的列向量，支持多列}

\item{suggestedColumns}{查看数据时推荐的显示列向量，支持多列，}

\item{titleColumn}{标题列，不支持多列}

\item{drop}{如果为TRUE，就先清理已有的数据集数据和定义}

\item{withData}{将样本数据写入数据集}

\item{desc}{对数据集的额外描述}
}
\description{
1、通过dsName和topic识别数据集和配置文件位置。

2、数据集必须首先使用\code{ds_init}函数创建配置文件（或手工建立），
然后才能对数据集做增删改查操作。

3、所有数据集的存储结构都使用@action作为第一层分区：
__APPEND__, __ARCHIVE__
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_as_deleted}()},
\code{\link{ds_as_from}()},
\code{\link{ds_collect}()},
\code{\link{ds_data}()},
\code{\link{ds_delete}()},
\code{\link{ds_drop}()},
\code{\link{ds_exists}()},
\code{\link{ds_files}()},
\code{\link{ds_read0}()},
\code{\link{ds_read_affected}()},
\code{\link{ds_read_appended}()},
\code{\link{ds_read_archived}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_parts}()},
\code{\link{ds_search}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
