% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-plyr.R
\name{ds_inner_join}
\alias{ds_inner_join}
\title{内连接}
\usage{
ds_inner_join(x, y, x_by, y_by = x_by, ...)
}
\description{
内连接
}
\seealso{
Other joins function: 
\code{\link{ds_anti_join}()},
\code{\link{ds_cbind}()},
\code{\link{ds_full_join}()},
\code{\link{ds_left_join}()},
\code{\link{ds_rbind}()},
\code{\link{ds_right_join}()},
\code{\link{ds_semi_join}()}
}
\concept{joins function}
