% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-meta.R
\name{ds_meta}
\alias{ds_meta}
\title{查看数据集元数据}
\usage{
ds_meta(dsName, topic = "CACHE")
}
\arguments{
\item{dsName}{数据集名称}

\item{topic}{主题域}
}
\description{
查看数据集元数据
}
\seealso{
Other metadata function: 
\code{\link{ds_as_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_meta_write}()},
\code{\link{ds_schema}()}
}
\concept{metadata function}
