% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-meta.R
\name{ds_meta_write}
\alias{ds_meta_write}
\title{写入数据集的Yaml配置}
\usage{
ds_meta_write(
  dsName,
  meta = c(),
  ex = c(),
  data = tibble(),
  topic = "CACHE",
  type = "__UNKNOWN__"
)
}
\arguments{
\item{dsName}{数据集名称}

\item{meta}{主元数据内容}

\item{ex}{扩展的元数据内容}

\item{data}{样本数据，用来分析数据结构}

\item{topic}{数据集保存的主题目录，默认为CACHE}

\item{type}{数据集类型}
}
\description{
支持补写元数据的配置项
}
\seealso{
Other metadata function: 
\code{\link{ds_as_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_meta}()},
\code{\link{ds_schema}()}
}
\concept{metadata function}
