% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-plyr.R
\name{ds_rbind}
\alias{ds_rbind}
\title{行合并}
\usage{
ds_rbind(dx = list())
}
\description{
行合并
}
\seealso{
Other joins function: 
\code{\link{ds_anti_join}()},
\code{\link{ds_cbind}()},
\code{\link{ds_full_join}()},
\code{\link{ds_inner_join}()},
\code{\link{ds_left_join}()},
\code{\link{ds_right_join}()},
\code{\link{ds_semi_join}()}
}
\concept{joins function}
