% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_read0}
\alias{ds_read0}
\title{读取数据集}
\usage{
ds_read0(dsName, topic = "CACHE", noDeleted = T, meta = c("@from"))
}
\arguments{
\item{dsName}{数据集名称}

\item{topic}{主题域}

\item{noDeleted}{不返回标记为删除的数据}
}
\description{
读取数据集
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_as_deleted}()},
\code{\link{ds_as_from}()},
\code{\link{ds_collect}()},
\code{\link{ds_data}()},
\code{\link{ds_delete}()},
\code{\link{ds_drop}()},
\code{\link{ds_exists}()},
\code{\link{ds_files}()},
\code{\link{ds_init}()},
\code{\link{ds_read_affected}()},
\code{\link{ds_read_appended}()},
\code{\link{ds_read_archived}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_parts}()},
\code{\link{ds_search}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
