% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-meta.R
\name{ds_schema}
\alias{ds_schema}
\title{查看数据集架构}
\usage{
ds_schema(ds, topic = "CACHE")
}
\arguments{
\item{ds}{数据框或数据集名称}
}
\description{
查看数据集架构时，必须包含数据。
}
\seealso{
Other metadata function: 
\code{\link{ds_as_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_meta_write}()},
\code{\link{ds_meta}()}
}
\concept{metadata function}
