% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-package.R
\docType{package}
\name{gali-package}
\alias{gali}
\alias{gali-package}
\title{gali: A Simpole Database for Analysis base on Apache Parquet}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: hongwei xue \email{43801@qq.com}

}
\keyword{internal}
