% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{获得当前配置}
\usage{
get_config()
}
\description{
获得当前配置
}
\seealso{
Other config functions: 
\code{\link{config_init}()},
\code{\link{config_load}()},
\code{\link{config_write}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{set_topic}()}
}
\concept{config functions}
