% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdays.R
\name{holidays_zh_CN}
\alias{holidays_zh_CN}
\title{节假日清单}
\usage{
holidays_zh_CN(fromDay = "2018-01-01", toDay = "2023-12-31")
}
\arguments{
\item{fromDay}{开始日期}

\item{toDay}{截止日期}
}
\description{
节假日清单
}
\seealso{
Other workday-funcs: 
\code{\link{ds_add_holidays}()},
\code{\link{ds_add_workdays}()},
\code{\link{holiday_assign}()},
\code{\link{workday_assign}()},
\code{\link{workdays_zh_CN}()}
}
\concept{workday-funcs}
