% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_gen}
\alias{id_gen}
\title{生成增量批次号}
\usage{
id_gen(prefix = "SN")
}
\description{
由于1秒内几乎不可能产生几十万个以上的批次，因此循环的seq和时间戳，可获得增量批次号
}
\seealso{
Other utils function: 
\code{\link{\%empty\%}()},
\code{\link{\%nin\%}()},
\code{\link{\%not-empty\%}()},
\code{\link{\%not-regex\%}()},
\code{\link{\%regex\%}()},
\code{\link{create_dir}()},
\code{\link{id_parse}()},
\code{\link{remove_dir}()}
}
\concept{utils function}
