% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_parse}
\alias{id_parse}
\title{解析批次号}
\usage{
id_parse(str)
}
\description{
解析批次号
}
\seealso{
Other utils function: 
\code{\link{\%empty\%}()},
\code{\link{\%nin\%}()},
\code{\link{\%not-empty\%}()},
\code{\link{create_dir}()},
\code{\link{id_gen}()},
\code{\link{remove_dir}()}
}
\concept{utils function}
