% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_changed}
\alias{import_changed}
\title{按批次导入文件素材}
\usage{
import_changed(
  importFolderFrom = "scheduleJob_2000-01-01",
  batchRegex = "scheduleJob_\\\\d{4}-\\\\d{2}-\\\\d{2}$",
  importTopic = "IMPORT"
)
}
\description{
按批次导入文件素材
}
\seealso{
Other import function: 
\code{\link{import_csv_preview}()},
\code{\link{import_csv}()},
\code{\link{import_files_preview}()},
\code{\link{import_files}()},
\code{\link{import_init}()},
\code{\link{import_scan}()},
\code{\link{import_search}()},
\code{\link{import_todo_flag}()}
}
\concept{import function}
