% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_scan}
\alias{import_scan}
\title{将需要导入的批次文件夹入库保存}
\usage{
import_scan(
  importFolderFrom = "scheduleJob_2000-01-01",
  batchRegex = "scheduleJob_\\\\d{4}-\\\\d{2}-\\\\d{2}$",
  importDataset = "__IMPORT_FILES__",
  importTopic = "IMPORT",
  cacheTopic = "CACHE"
)
}
\description{
比较根文件夹下的所有子文件夹比较，扫描所有比importFolderFrom字符串小的
}
\seealso{
Other import function: 
\code{\link{import_changed}()},
\code{\link{import_csv_preview}()},
\code{\link{import_csv}()},
\code{\link{import_files_preview}()},
\code{\link{import_files}()},
\code{\link{import_init}()},
\code{\link{import_search}()},
\code{\link{import_todo_flag}()}
}
\concept{import function}
