% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_filedType}
\alias{schema_filedType}
\title{自动识别R对象的FiledType}
\usage{
schema_filedType(val)
}
\description{
自动识别R对象的FiledType
}
\seealso{
Other schema-functions: 
\code{\link{schema_check}()},
\code{\link{schema_from_meta}()},
\code{\link{schema_from_tibble}()},
\code{\link{schema_name}()},
\code{\link{schema_to_arrow}()},
\code{\link{schema_to_meta}()},
\code{\link{schema_to_tibble}()},
\code{\link{schema_types}()}
}
\concept{schema-functions}
