% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_from_tibble}
\alias{schema_from_tibble}
\title{从数据框的R类型转换schema}
\usage{
schema_from_tibble(d)
}
\description{
tibble data >> schema list

根据数据框获得fieldName，进一步推断rType，以及fieldType、asTibble函数和asArrow函数
默认匹配shinyType

仅支持可写入数据集的类型转换，如果包括无法识别的列表、S4等类型将会报错
}
\seealso{
Other schema-functions: 
\code{\link{schema_check}()},
\code{\link{schema_filedType}()},
\code{\link{schema_from_meta}()},
\code{\link{schema_name}()},
\code{\link{schema_to_arrow}()},
\code{\link{schema_to_meta}()},
\code{\link{schema_to_tibble}()},
\code{\link{schema_types}()}
}
\concept{schema-functions}
