% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_name}
\alias{schema_name}
\title{检查数据类型}
\usage{
schema_name(d = NULL)
}
\arguments{
\item{d}{用于检查的数据框，为空时罗列支持的字段类型}
}
\description{
检查数据类型
}
\seealso{
Other schema-functions: 
\code{\link{schema_check}()},
\code{\link{schema_filedType}()},
\code{\link{schema_from_meta}()},
\code{\link{schema_from_tibble}()},
\code{\link{schema_to_arrow}()},
\code{\link{schema_to_meta}()},
\code{\link{schema_to_tibble}()},
\code{\link{schema_types}()}
}
\concept{schema-functions}
