% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_to_arrow}
\alias{schema_to_arrow}
\title{将schema转换为arrow要求的对象}
\usage{
schema_to_arrow(bySchema = tibble())
}
\description{
schema list >> arrow object

读取arrow时
}
\seealso{
Other schema-functions: 
\code{\link{schema_check}()},
\code{\link{schema_filedType}()},
\code{\link{schema_from_meta}()},
\code{\link{schema_from_tibble}()},
\code{\link{schema_name}()},
\code{\link{schema_to_meta}()},
\code{\link{schema_to_tibble}()},
\code{\link{schema_types}()}
}
\concept{schema-functions}
