% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_to_meta}
\alias{schema_to_meta}
\title{从chema数据框生成元数据}
\usage{
schema_to_meta(bySchema = tibble::tibble())
}
\arguments{
\item{bySchema}{提供fieldName和fieldType}
}
\description{
从chema数据框生成元数据
}
\seealso{
Other schema-functions: 
\code{\link{schema_check}()},
\code{\link{schema_filedType}()},
\code{\link{schema_from_meta}()},
\code{\link{schema_from_tibble}()},
\code{\link{schema_name}()},
\code{\link{schema_to_arrow}()},
\code{\link{schema_to_tibble}()},
\code{\link{schema_types}()}
}
\concept{schema-functions}
