% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_to_tibble}
\alias{schema_to_tibble}
\title{按照shcema转换数据框}
\usage{
schema_to_tibble(d, bySchema = list())
}
\description{
schema list >> tibble data
}
\seealso{
Other schema-functions: 
\code{\link{schema_check}()},
\code{\link{schema_filedType}()},
\code{\link{schema_from_meta}()},
\code{\link{schema_from_tibble}()},
\code{\link{schema_name}()},
\code{\link{schema_to_arrow}()},
\code{\link{schema_to_meta}()},
\code{\link{schema_types}()}
}
\concept{schema-functions}
