% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_types}
\alias{schema_types}
\title{所有数据类型}
\usage{
schema_types()
}
\description{
所有数据类型
}
\seealso{
Other schema-functions: 
\code{\link{schema_check}()},
\code{\link{schema_filedType}()},
\code{\link{schema_from_meta}()},
\code{\link{schema_from_tibble}()},
\code{\link{schema_name}()},
\code{\link{schema_to_arrow}()},
\code{\link{schema_to_meta}()},
\code{\link{schema_to_tibble}()}
}
\concept{schema-functions}
