% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_topic}
\alias{set_topic}
\title{设置配置项的值}
\usage{
set_topic(topic, path)
}
\description{
设置配置项的值
}
\seealso{
Other config functions: 
\code{\link{config_init}()},
\code{\link{config_load}()},
\code{\link{config_write}()},
\code{\link{get_config}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()}
}
\concept{config functions}
