% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-task.R
\name{task_load}
\alias{task_load}
\title{加载RMD文件为多个段落}
\usage{
task_load(rmdName, topic = "TASK")
}
\description{
加载RMD文件为多个段落
}
\seealso{
Other rmd scripts functions: 
\code{\link{task_all}()},
\code{\link{task_path}()},
\code{\link{task_render}()},
\code{\link{task_run}()},
\code{\link{task_save}()},
\code{\link{task_yaml}()}
}
\concept{rmd scripts functions}
