% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-task.R
\name{task_path}
\alias{task_path}
\title{RMD脚本路径}
\usage{
task_path(rmdName, type = c("rmd", "output", "script"), topic = "TASK")
}
\description{
RMD脚本路径
}
\seealso{
Other rmd scripts functions: 
\code{\link{task_all}()},
\code{\link{task_load}()},
\code{\link{task_render}()},
\code{\link{task_run}()},
\code{\link{task_save}()},
\code{\link{task_yaml}()}
}
\concept{rmd scripts functions}
