% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-task.R
\name{task_render}
\alias{task_render}
\title{直接运行RMD文件包含的有效脚本}
\usage{
task_render(rmdName, topic = "TASK", pure = FALSE, quiet = TRUE)
}
\description{
直接运行RMD文件包含的有效脚本
}
\seealso{
Other rmd scripts functions: 
\code{\link{task_all}()},
\code{\link{task_load}()},
\code{\link{task_path}()},
\code{\link{task_run}()},
\code{\link{task_save}()},
\code{\link{task_yaml}()}
}
\concept{rmd scripts functions}
