% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-task.R
\name{task_save}
\alias{task_save}
\title{保存RMD脚本}
\usage{
task_save(newParams = list(), rmdName, topic = "TASK")
}
\description{
保存RMD脚本
}
\seealso{
Other rmd scripts functions: 
\code{\link{task_all}()},
\code{\link{task_load}()},
\code{\link{task_path}()},
\code{\link{task_render}()},
\code{\link{task_run}()},
\code{\link{task_yaml}()}
}
\concept{rmd scripts functions}
