% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-task.R
\name{task_yaml}
\alias{task_yaml}
\title{加载RMD文件的Yaml元数据}
\usage{
task_yaml(rmdName = NULL, contents = "", topic = "TASK")
}
\description{
加载RMD文件的Yaml元数据
}
\seealso{
Other rmd scripts functions: 
\code{\link{task_all}()},
\code{\link{task_load}()},
\code{\link{task_path}()},
\code{\link{task_render}()},
\code{\link{task_run}()},
\code{\link{task_save}()}
}
\concept{rmd scripts functions}
