% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdays.R
\name{workday_assign}
\alias{workday_assign}
\title{增加工作日调整}
\usage{
workday_assign(from, to = from)
}
\arguments{
\item{fromDay}{开始日期}

\item{toDay}{截止日期}
}
\description{
增加工作日调整
}
\seealso{
Other workday-funcs: 
\code{\link{ds_add_holidays}()},
\code{\link{ds_add_workdays}()},
\code{\link{holiday_assign}()},
\code{\link{holidays_zh_CN}()},
\code{\link{workdays_zh_CN}()}
}
\concept{workday-funcs}
