% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdays.R
\name{workdays_zh_CN}
\alias{workdays_zh_CN}
\title{工作日清单}
\usage{
workdays_zh_CN(fromDay = "2018-01-01", toDay = "2023-12-31")
}
\arguments{
\item{fromDay}{开始日期}

\item{toDay}{截止日期}
}
\description{
默认为周一至周五为工作日，再根据中国国务院发布的工作日调整政策修正
}
\seealso{
Other workday-funcs: 
\code{\link{ds_add_holidays}()},
\code{\link{ds_add_workdays}()},
\code{\link{holiday_assign}()},
\code{\link{holidays_zh_CN}()},
\code{\link{workday_assign}()}
}
\concept{workday-funcs}
